% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{call.to.char}
\alias{call.to.char}
\title{Convert calls to character}
\usage{
call.to.char(x)
}
\arguments{
\item{x}{a formula (call)}
}
\value{
a character vector of length 1
}
\description{
Utility function (hack) to convert calls such as y~x to their character
equivalent
}
\details{
It would be nice if \code{as.character(y~x)} gave "y~x", but it doesn't, so
this hack achieves the same goal
}
\examples{

as.character(y~x)
call.to.char(y~x)

}
\author{
Ben Bolker
}
\keyword{misc}

