% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{get.mnames}
\alias{get.mnames}
\title{extract model names}
\usage{
get.mnames(Call)
}
\arguments{
\item{Call}{a function call (usually a list of models)}
}
\value{
a vector of model names
}
\description{
given a list of models, extract the names (or "model n")
}
\author{
Ben Bolker
}
\keyword{misc}

