% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle2.options}
\alias{mle2.options}
\title{Options for maximum likelihood estimation}
\usage{
mle2.options(...)
}
\arguments{
\item{\dots}{names of arguments to query, or a list of values to set}
}
\value{
Values of queried parameters, or (invisibly) the full list of
parameters
}
\description{
Query or set MLE parameters
}
\details{
\itemize{
\item optim.method name of optimization method (see
\code{\link{optim}} for choices)
\item confintn ame of confidence-interval:
choices are "spline", "uniroot", "hessian" corresponding to spline
inversion, attempt to find best answer via uniroot, information-matrix
approximation
\item optimizer optimization function to use by default
(choices: "optim", "nlm", "nlminb", "constrOptim")
}
}
\seealso{
\code{\link{mle2-class}}
}
\keyword{models}

