% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impsamp.R
\name{pop_pred_samp}
\alias{pop_pred_samp}
\title{generate population prediction sample from parameters (EXPERIMENTAL)}
\usage{
pop_pred_samp(object, n = 1000, n_imp = n * 10, return_wts = FALSE,
  impsamp = FALSE, PDify = FALSE, PDmethod=NULL, tol = 1e-06,
  fix_params = NULL)
}
\arguments{
\item{object}{a fitted \code{mle2} object}

\item{n}{number of samples to return}

\item{n_imp}{number of total samples from which to draw, if doing importance sampling}

\item{return_wts}{return a column giving the weights of the samples, for use in weighted summaries?}

\item{impsamp}{subsample values (with replacement) based on their weights?}

\item{PDify}{try to correct non-positive-definite matrix if necessary? (EXPERIMENTAL)}
\item{PDmethod}{method to use for correcting non-pos-def matrices: the
  default NULL will use the King et al generalized inverse/Cholesky
  approach for semidefinite matrices and \code{Matrix::nearPD} for
  matrices with definitely negative eigenvalues}
\item{tol}{tolerance for detecting small eigenvalues}
\item{fix_params}{parameters to fix (in addition to parameters that were fixed during estimation)}
}
\note{Turn on non-positive-definite corrections (i.e. set
  \code{PDify=TRUE}) with caution!}
\description{
generate population prediction sample from parameters
}
\references{
Gill, Jeff, and Gary King. "What to Do When Your Hessian Is Not Invertible: Alternatives to Model Respecification in Nonlinear Estimation." Sociological Methods & Research 33, no. 1 (2004): 54-87.
}
