\name{get.mnames}
\alias{get.mnames}
\title{extract model names}
\description{
  given a list of models, extract the names (or "model n")
}
\usage{
get.mnames(Call)
}
\arguments{
  \item{Call}{a function call (usually a list of models)}
}
\value{
  a vector of model names
}
\author{Ben Bolker}
\keyword{misc}
