% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impsamp.R
\name{pop_pred_samp}
\alias{pop_pred_samp}
\title{generate population prediction sample from parameters}
\usage{
pop_pred_samp(object, n = 1000, n_imp = n * 10, return_wts = FALSE,
  impsamp = FALSE, ginv = FALSE, tol = 1e-06)
}
\arguments{
\item{object}{a fitted \code{mle2} object}

\item{n}{number of samples to return}

\item{n_imp}{number of total samples from which to draw, if doing importance sampling}

\item{return_wts}{return a column giving the weights of the samples, for use in weighted summaries?}

\item{impsamp}{subsample values (with replacement) based on their weights?}

\item{ginv}{use Gill and King generalized-inverse procedure to correct non-positive-definite variance-covariance matrix if necessary?}

\item{tol}{tolerance for detecting small eigenvalues}
}
\description{
generate population prediction sample from parameters
}
\references{
Gill, Jeff, and Gary King. "What to Do When Your Hessian Is Not Invertible: Alternatives to Model Respecification in Nonlinear Estimation." Sociological Methods & Research 33, no. 1 (2004): 54-87.
}
