\name{slice}
\alias{slice}
\title{Calculate likelihood "slices"}
\description{
  Computes a cross-section of a multi-dimensional likelihood surface
}
\usage{
slice(fitted, ...)
}
\arguments{
  \item{fitted}{a fitted object of class \code{mle}}
  \item{\dots}{other arguments: which etc.}
}
\value{
  still a stub.
}
\note{
  Slices provide a lighter-weight way to explore likelihood surfaces
  than profiles, since they vary a single parameter rather than 
}
\author{Ben Bolker}
\seealso{\code{\link{profile}}}
\keyword{misc}
